adding single param vs param plot
